move browser profile select to new component
