+++ frontend/src/pages/archive/crawl-templates-new.ts
@@ -3,7 +3,6 @@ import { ifDefined } from "lit/directives/if-defined.js";
 import { msg, localized, str } from "@lit/localize";
 import cronParser from "cron-parser";
 import { parse as yamlToJson, stringify as jsonToYaml } from "yaml";
-import orderBy from "lodash/fp/orderBy";
 
 import type { AuthState } from "../../utils/AuthService";
 import LiteElement, { html } from "../../utils/LiteElement";
@@ -91,10 +90,7 @@ export class CrawlTemplatesNew extends LiteElement {
   private serverError?: string;
 
   @state()
-  browserProfiles?: Profile[];
-
-  @state()
-  selectedProfile?: Profile;
+  private selectedProfile?: Profile;
 
   private get timeZone() {
     return Intl.DateTimeFormat().resolvedOptions().timeZone;
@@ -147,10 +143,6 @@ export class CrawlTemplatesNew extends LiteElement {
     };
     this.configCode = jsonToYaml(this.initialCrawlTemplate.config);
     super.connectedCallback();
-  }
-
-  protected firstUpdated() {
-    this.fetchBrowserProfiles();
   }
 
   render() {
@@ -250,91 +242,12 @@ export class CrawlTemplatesNew extends LiteElement {
         ></sl-input>
 
         <div>
-          <sl-select
-            label=${msg("Browser Profile")}
-            clearable
-            value=${this.selectedProfile?.id || ""}
-            placeholder=${this.browserProfiles
-              ? msg("Select Profile")
-              : msg("Loading")}
-            ?disabled=${!this.browserProfiles?.length}
-            @sl-change=${(e: any) =>
-              (this.selectedProfile = this.browserProfiles?.find(
-                ({ id }) => id === e.target.value
-              ))}
-            @sl-focus=${() => {
-              // Refetch to keep list up to date
-              this.fetchBrowserProfiles();
-            }}
-          >
-            ${this.browserProfiles
-              ? ""
-              : html` <sl-spinner slot="prefix"></sl-spinner> `}
-            ${this.browserProfiles?.map(
-              (profile) => html`
-                <sl-menu-item value=${profile.id}>
-                  ${profile.name}
-                  <div slot="suffix">
-                    <div class="text-xs">
-                      <sl-format-date
-                        date=${`${profile.created}Z` /** Z for UTC */}
-                        month="2-digit"
-                        day="2-digit"
-                        year="2-digit"
-                      ></sl-format-date>
-                    </div></div
-                ></sl-menu-item>
-              `
-            )}
-          </sl-select>
-
-          ${this.browserProfiles && !this.browserProfiles.length
-            ? html`
-                <div class="mt-2 text-sm text-neutral-500">
-                  <span class="inline-block align-middle"
-                    >${msg("No browser profiles found.")}</span
-                  >
-                  <a
-                    href=${`/archives/${this.archiveId}/browser-profiles/new`}
-                    class="font-medium text-primary hover:text-indigo-500"
-                    target="_blank"
-                    ><span class="inline-block align-middle"
-                      >${msg("Create a browser profile")}</span
-                    >
-                    <sl-icon
-                      class="inline-block align-middle"
-                      name="box-arrow-up-right"
-                    ></sl-icon
-                  ></a>
-                </div>
-              `
-            : ""}
-          ${this.selectedProfile
-            ? html`
-                <div
-                  class="mt-2 border bg-slate-50 border-slate-100 rounded p-2 text-sm flex justify-between"
-                >
-                  ${this.selectedProfile.description
-                    ? html`<em class="text-slate-500"
-                        >${this.selectedProfile.description}</em
-                      >`
-                    : ""}
-                  <a
-                    href=${`/archives/${this.archiveId}/browser-profiles/profile/${this.selectedProfile.id}`}
-                    class="font-medium text-primary hover:text-indigo-500"
-                    target="_blank"
-                  >
-                    <span class="inline-block align-middle mr-1"
-                      >${msg("View profile")}</span
-                    >
-                    <sl-icon
-                      class="inline-block align-middle"
-                      name="box-arrow-up-right"
-                    ></sl-icon>
-                  </a>
-                </div>
-              `
-            : ""}
+          <btrix-select-browser-profile
+            archiveId=${this.archiveId}
+            initialProfileId=${this.initialCrawlTemplate?.profileid || ""}
+            .authState=${this.authState}
+            .selectedProfile=${this.selectedProfile}
+          ></btrix-select-browser-profile>
         </div>
       </section>
     `;
@@ -666,40 +579,6 @@ export class CrawlTemplatesNew extends LiteElement {
       minute,
       period,
     });
-  }
-
-  /**
-   * Fetch browser profiles and update internal state
-   */
-  private async fetchBrowserProfiles(): Promise<void> {
-    try {
-      const data = await this.getProfiles();
-
-      this.browserProfiles = orderBy(["name", "created"])(["asc", "desc"])(
-        data
-      ) as Profile[];
-
-      if (this.initialCrawlTemplate?.profileid && !this.selectedProfile) {
-        this.selectedProfile = this.browserProfiles.find(
-          ({ id }) => id === this.initialCrawlTemplate?.profileid
-        );
-      }
-    } catch (e) {
-      this.notify({
-        message: msg("Sorry, couldn't retrieve browser profiles at this time."),
-        type: "danger",
-        icon: "exclamation-octagon",
-      });
-    }
-  }
-
-  private async getProfiles(): Promise<Profile[]> {
-    const data = await this.apiFetch(
-      `/archives/${this.archiveId}/profiles`,
-      this.authState!
-    );
-
-    return data;
   }
 }
